requirements.txt
colorthief
flask
nfcpy
pyaudio
pydub
pyusb
pyyaml
simpleaudio
soundfile
tekore
